refactor(profile): extract self-contained widgets from profile_setup_screen (#4512)
First commit of the profile_setup_screen.dart decomposition (epic #4339).
Behavior-preserving, UI-only move — all logic stays in ProfileEditorBloc /
MyProfileBloc; no backend, bloc, or repository changes.

- Create screens/profile_setup/ (view/ + widgets/ + barrels) mirroring the
  screens/search_results/ convention.
- Move the Page + View into view/. Keep the public ProfileSetupScreen class,
  its route consts (editPath/setupPath/editRouteName/setupRouteName), and the
  isNewUser flag so every routing call-site is unchanged.
- Extract the already self-contained widgets to widgets/: SaveButton,
  BannerEditingBlock (+ preview/action/swatch), PublicKeyLink,
  VerifiedAccountsSection (+ GetVerifiedTile), and the upload-error message
  mappers. launchVerifierFlow stays with the View (its only caller;
  @visibleForTesting).
- Repoint the six importers to the new barrel; make the l10n hardcoded-string
  guard read the whole profile_setup/ directory.

The View still holds the inline form (avatar/name/bio/website/username);
extracting those sections lands in the next commit.
