+++ mobile/lib/screens/profile_setup/profile_setup.dart
@@ -1,0 +1,2 @@
+export 'view/view.dart';
+export 'widgets/widgets.dart';
