+++ mobile/lib/screens/profile_setup/view/profile_setup_view.dart
@@ -1,6 +1,3 @@
-// ABOUTME: Profile setup screen for new users to configure their display name, bio, and avatar
-// ABOUTME: Publishes initial profile metadata to Nostr after setup is complete
-
 import 'dart:async';
 import 'dart:io';
 
@@ -22,117 +19,12 @@ import 'package:openvine/providers/app_providers.dart';
 import 'package:openvine/providers/nostr_client_provider.dart';
 import 'package:openvine/providers/user_profile_providers.dart';
 import 'package:openvine/screens/apps/web_iframe_sandbox_screen.dart';
-import 'package:openvine/screens/key_management_screen.dart';
-import 'package:openvine/widgets/branded_loading_scaffold.dart';
+import 'package:openvine/screens/profile_setup/widgets/widgets.dart';
 import 'package:openvine/widgets/profile/nostr_info_sheet_content.dart';
-import 'package:openvine/widgets/profile/verified_accounts_row.dart';
 import 'package:openvine/widgets/profile_editor/username_status_indicator.dart';
 import 'package:openvine/widgets/user_avatar.dart';
-import 'package:openvine/widgets/vine_cached_image.dart';
-import 'package:profile_repository/profile_repository.dart';
 import 'package:unified_logger/unified_logger.dart';
 import 'package:url_launcher/url_launcher.dart';
-
-/// Maps an [AvatarUploadError] case to its localized snackbar string.
-///
-/// The bloc classifies upload failures; the UI just picks the right l10n key.
-/// Keeping this here colocates UI copy with the screen that shows it.
-String profileSetupUploadErrorMessage(
-  AppLocalizations l10n,
-  AvatarUploadError error,
-) {
-  return switch (error) {
-    AvatarUploadError.network => l10n.profileSetupUploadNetworkError,
-    AvatarUploadError.auth => l10n.profileSetupUploadAuthError,
-    AvatarUploadError.fileTooLarge => l10n.profileSetupUploadFileTooLarge,
-    AvatarUploadError.server => l10n.profileSetupUploadServerError,
-    AvatarUploadError.generic => l10n.profileSetupUploadFailedGeneric,
-  };
-}
-
-/// Maps a [BannerUploadError] case to its localized snackbar string.
-///
-/// Reuses the same upload-error copy as the avatar — the failure modes
-/// are identical from the user's point of view.
-String profileSetupBannerUploadErrorMessage(
-  AppLocalizations l10n,
-  BannerUploadError error,
-) {
-  return switch (error) {
-    BannerUploadError.network => l10n.profileSetupUploadNetworkError,
-    BannerUploadError.auth => l10n.profileSetupUploadAuthError,
-    BannerUploadError.fileTooLarge => l10n.profileSetupUploadFileTooLarge,
-    BannerUploadError.server => l10n.profileSetupUploadServerError,
-    BannerUploadError.generic => l10n.profileSetupUploadFailedGeneric,
-  };
-}
-
-class ProfileSetupScreen extends ConsumerWidget {
-  /// Route name for editing existing profile.
-  static const editRouteName = 'edit-profile';
-
-  /// Path for editing existing profile.
-  static const editPath = '/edit-profile';
-
-  /// Route name for setting up new profile.
-  static const setupRouteName = 'setup-profile';
-
-  /// Path for setting up new profile.
-  static const setupPath = '/setup-profile';
-
-  const ProfileSetupScreen({required this.isNewUser, super.key});
-
-  final bool isNewUser;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final profileRepository = ref.watch(profileRepositoryProvider);
-    final blossomUploadService = ref.watch(blossomUploadServiceProvider);
-    final authService = ref.watch(authServiceProvider);
-    final identityClaimsRepository = ref.watch(
-      identityClaimsRepositoryProvider,
-    );
-
-    final pubkey = authService.currentPublicKeyHex;
-
-    // Show loading until NostrClient has keys
-    if (profileRepository == null || pubkey == null) {
-      return const BrandedLoadingScaffold();
-    }
-
-    return MultiBlocProvider(
-      providers: [
-        BlocProvider<ProfileEditorBloc>(
-          // Riverpod-into-BlocProvider bridge: rebuild the bloc when any
-          // captured dependency's identity changes (auth flip, account
-          // switch). Record-typed key compares per-field with `==`; the
-          // captured classes do not override `==`, so equality falls
-          // through to identity — exactly the semantics this needs.
-          // See `state_management.md`.
-          key: ValueKey((profileRepository, blossomUploadService, pubkey)),
-          create: (context) => ProfileEditorBloc(
-            profileRepository: profileRepository,
-            blossomUploadService: blossomUploadService,
-            hasExistingProfile: authService.hasExistingProfile,
-            currentUserPubkey: pubkey,
-          ),
-        ),
-        BlocProvider<MyProfileBloc>(
-          create: (context) {
-            final bloc = MyProfileBloc(
-              profileRepository: profileRepository,
-              pubkey: pubkey,
-              identityClaimsRepository: identityClaimsRepository,
-            );
-            if (!isNewUser) bloc.add(const MyProfileLoadRequested());
-            return bloc;
-          },
-        ),
-      ],
-      child: ProfileSetupScreenView(isNewUser: isNewUser),
-    );
-  }
-}
 
 class ProfileSetupScreenView extends ConsumerStatefulWidget {
   const ProfileSetupScreenView({required this.isNewUser, super.key});
@@ -920,9 +812,9 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
                               ),
                               const SizedBox(height: 16),
 
-                              const _VerifiedAccountsSection(),
+                              const VerifiedAccountsSection(),
 
-                              const _PublicKeyLink(),
+                              const PublicKeyLink(),
                               const SizedBox(height: 16),
 
                               // NIP-05 Username (optional)
@@ -1046,7 +938,7 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
                               // Replaces the old standalone profile-color
                               // picker; the bloc serializes the chosen color
                               // into the same kind-0 `banner` field.
-                              const _BannerEditingBlock(),
+                              const BannerEditingBlock(),
                             ],
                           ),
                         ),
@@ -1106,7 +998,7 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
                     const SizedBox(width: 16),
                     if (pubkey != null)
                       Expanded(
-                        child: _SaveButton(
+                        child: SaveButton(
                           canSave: profileEditorState.isSaveReady,
                           onSave: () {
                             if (_nameController.text.trim().isEmpty) {
@@ -1460,417 +1352,6 @@ class _ProfileSetupScreenViewState extends ConsumerState<ProfileSetupScreenView>
   }
 }
 
-class _SaveButton extends StatelessWidget {
-  const _SaveButton({required this.canSave, required this.onSave});
-
-  final bool canSave;
-  final VoidCallback onSave;
-
-  @override
-  Widget build(BuildContext context) {
-    final isLoading = context.select<ProfileEditorBloc, bool>(
-      (bloc) => bloc.state.status == ProfileEditorStatus.loading,
-    );
-
-    return ElevatedButton(
-      onPressed: (isLoading || !canSave) ? null : onSave,
-      style: ElevatedButton.styleFrom(
-        backgroundColor: VineTheme.vineGreen,
-        foregroundColor: VineTheme.onPrimary,
-        disabledBackgroundColor: VineTheme.vineGreen.withValues(alpha: 0.4),
-        disabledForegroundColor: VineTheme.onPrimary.withValues(alpha: 0.6),
-        padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 16),
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-      ),
-      child: isLoading
-          ? Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                const SizedBox(
-                  width: 20,
-                  height: 20,
-                  child: CircularProgressIndicator(
-                    strokeWidth: 2,
-                    color: VineTheme.onPrimary,
-                  ),
-                ),
-                const SizedBox(width: 12),
-                Text(
-                  context.l10n.profileSetupSavingButton,
-                  style: VineTheme.titleMediumFont(color: VineTheme.onPrimary),
-                ),
-              ],
-            )
-          : Text(
-              context.l10n.profileSetupSaveButton,
-              style: VineTheme.titleMediumFont(color: VineTheme.onPrimary),
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-            ),
-    );
-  }
-}
-
-/// Banner editing block: 3:1 preview, gallery upload, and color swatches.
-///
-/// Image and color are mutually exclusive — selecting one clears the other
-/// at the bloc layer. The preview reads
-/// `pendingBannerUrl ?? pendingBannerColor ?? persistedBanner` via
-/// granular `context.select`s so unrelated state changes don't rebuild it.
-class _BannerEditingBlock extends StatelessWidget {
-  const _BannerEditingBlock();
-
-  // Brand-accent palette parallel to the avatar's profile-color picker.
-  // Order is load-bearing for the
-  // `profile_banner_color_swatch_preset_<index>` keys used in tests.
-  static const List<Color> _bannerSwatchPalette = [
-    VineTheme.vineGreen,
-    VineTheme.accentBlue,
-    VineTheme.accentPurple,
-    VineTheme.likeRed,
-    VineTheme.accentOrange,
-    VineTheme.accentLime,
-    VineTheme.accentPink,
-    VineTheme.accentViolet,
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = context.l10n;
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(bottom: 8),
-            child: Text(
-              l10n.profileSetupBannerSectionTitle,
-              style: VineTheme.labelMediumFont(
-                color: VineTheme.onSurfaceMuted,
-              ),
-            ),
-          ),
-          const _BannerPreview(),
-          const SizedBox(height: 12),
-          const _BannerActionRow(),
-          const SizedBox(height: 12),
-          Wrap(
-            spacing: 12,
-            runSpacing: 12,
-            children: [
-              for (var i = 0; i < _bannerSwatchPalette.length; i++)
-                _BannerColorSwatch(
-                  index: i,
-                  color: _bannerSwatchPalette[i],
-                ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _BannerPreview extends StatelessWidget {
-  const _BannerPreview();
-
-  @override
-  Widget build(BuildContext context) {
-    final pendingUrl = context.select(
-      (ProfileEditorBloc b) => b.state.pendingBannerUrl,
-    );
-    final pendingColor = context.select(
-      (ProfileEditorBloc b) => b.state.pendingBannerColor,
-    );
-    final persistedBanner = context.select(
-      (ProfileEditorBloc b) => b.state.persistedBanner,
-    );
-    final isUploading = context.select(
-      (ProfileEditorBloc b) =>
-          b.state.pendingBannerStatus == PendingBannerStatus.uploading,
-    );
-
-    final radius = BorderRadius.circular(16);
-    final imageUrl = (pendingUrl != null && pendingUrl.isNotEmpty)
-        ? pendingUrl
-        : (pendingColor == null &&
-              persistedBanner != null &&
-              persistedBanner.startsWith('http'))
-        ? persistedBanner
-        : null;
-
-    Widget child;
-    Key previewKey;
-    if (imageUrl != null) {
-      previewKey = const ValueKey('profile_banner_image_preview');
-      child = Semantics(
-        label: context.l10n.profileSetupBannerSectionTitle,
-        image: true,
-        child: ClipRRect(
-          borderRadius: radius,
-          child: VineCachedImage(imageUrl: imageUrl),
-        ),
-      );
-    } else if (pendingColor != null) {
-      previewKey = const ValueKey('profile_banner_color_preview');
-      child = DecoratedBox(
-        decoration: BoxDecoration(
-          color: pendingColor,
-          borderRadius: radius,
-        ),
-      );
-    } else {
-      previewKey = const ValueKey('profile_banner_empty_preview');
-      child = DecoratedBox(
-        decoration: BoxDecoration(
-          color: VineTheme.surfaceContainer,
-          borderRadius: radius,
-          border: Border.all(color: VineTheme.outlineMuted, width: 2),
-        ),
-      );
-    }
-
-    return Stack(
-      children: [
-        AspectRatio(
-          key: previewKey,
-          aspectRatio: 3,
-          child: child,
-        ),
-        if (isUploading)
-          Positioned.fill(
-            child: DecoratedBox(
-              decoration: BoxDecoration(
-                color: VineTheme.backgroundColor.withValues(alpha: 0.6),
-                borderRadius: radius,
-              ),
-              child: const Center(
-                child: CircularProgressIndicator(
-                  color: VineTheme.vineGreen,
-                  strokeWidth: 3,
-                ),
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-}
-
-class _BannerActionRow extends StatelessWidget {
-  const _BannerActionRow();
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = context.l10n;
-    final hasSelection = context.select(
-      (ProfileEditorBloc b) =>
-          (b.state.pendingBannerUrl?.isNotEmpty ?? false) ||
-          b.state.pendingBannerColor != null ||
-          (b.state.persistedBanner?.isNotEmpty ?? false),
-    );
-    final isUploading = context.select(
-      (ProfileEditorBloc b) =>
-          b.state.pendingBannerStatus == PendingBannerStatus.uploading,
-    );
-
-    return Row(
-      children: [
-        Expanded(
-          child: OutlinedButton(
-            onPressed: isUploading ? null : () => _pickBannerImage(context),
-            style: OutlinedButton.styleFrom(
-              foregroundColor: VineTheme.vineGreen,
-              side: const BorderSide(
-                color: VineTheme.outlineMuted,
-                width: 2,
-              ),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(20),
-              ),
-              padding: const EdgeInsets.symmetric(vertical: 12),
-            ),
-            child: Text(
-              l10n.profileSetupBannerUploadButton,
-              style: VineTheme.titleMediumFont(color: VineTheme.vineGreen),
-            ),
-          ),
-        ),
-        const SizedBox(width: 12),
-        Expanded(
-          child: OutlinedButton(
-            onPressed: (isUploading || !hasSelection)
-                ? null
-                : () => context.read<ProfileEditorBloc>().add(
-                    const ProfileBannerCleared(),
-                  ),
-            style: OutlinedButton.styleFrom(
-              foregroundColor: VineTheme.lightText,
-              side: const BorderSide(
-                color: VineTheme.outlineMuted,
-                width: 2,
-              ),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(20),
-              ),
-              padding: const EdgeInsets.symmetric(vertical: 12),
-            ),
-            child: Text(
-              l10n.profileSetupBannerClearButton,
-              style: VineTheme.titleMediumFont(color: VineTheme.lightText),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Future<void> _pickBannerImage(BuildContext context) async {
-    final editorBloc = context.read<ProfileEditorBloc>();
-    final container = ProviderScope.containerOf(context, listen: false);
-    final pk = container.read(authServiceProvider).currentPublicKeyHex;
-    if (pk == null) return;
-
-    final picker = ImagePicker();
-    XFile? picked;
-    try {
-      picked = await picker.pickImage(
-        source: ImageSource.gallery,
-        maxWidth: 1500,
-        imageQuality: 85,
-        requestFullMetadata: false,
-      );
-    } catch (e) {
-      Log.error(
-        'Banner image_picker error: $e',
-        name: 'ProfileSetupScreen',
-        category: LogCategory.ui,
-      );
-      return;
-    }
-    if (picked == null) return;
-
-    if (kIsWeb) {
-      final bytes = await picked.readAsBytes();
-      editorBloc.add(
-        ProfileBannerUploadRequested(
-          pubkey: pk,
-          bytes: bytes,
-          filename: picked.name,
-        ),
-      );
-    } else {
-      editorBloc.add(
-        ProfileBannerUploadRequested(
-          pubkey: pk,
-          file: File(picked.path),
-        ),
-      );
-    }
-  }
-}
-
-class _BannerColorSwatch extends StatelessWidget {
-  const _BannerColorSwatch({required this.index, required this.color});
-
-  final int index;
-  final Color color;
-
-  @override
-  Widget build(BuildContext context) {
-    final isSelected = context.select(
-      (ProfileEditorBloc b) => b.state.pendingBannerColor == color,
-    );
-    return Semantics(
-      button: true,
-      label: context.l10n.profileSetupBannerSectionTitle,
-      child: GestureDetector(
-        key: ValueKey('profile_banner_color_swatch_preset_$index'),
-        onTap: () => context.read<ProfileEditorBloc>().add(
-          ProfileBannerColorSelected(color),
-        ),
-        child: Container(
-          width: 44,
-          height: 44,
-          decoration: BoxDecoration(
-            color: color,
-            borderRadius: BorderRadius.circular(12),
-            border: Border.all(
-              color: isSelected ? VineTheme.whiteText : VineTheme.transparent,
-              width: 3,
-            ),
-          ),
-          child: isSelected
-              ? const DivineIcon(
-                  icon: DivineIconName.check,
-                  color: VineTheme.whiteText,
-                  size: 20,
-                )
-              : null,
-        ),
-      ),
-    );
-  }
-}
-
-class _PublicKeyLink extends StatelessWidget {
-  const _PublicKeyLink();
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = context.l10n;
-    return Align(
-      alignment: AlignmentDirectional.centerStart,
-      child: TextButton(
-        onPressed: () => context.pushNamed(KeyManagementScreen.routeName),
-        child: Text(
-          l10n.profileEditPublicKeyLink,
-          style: VineTheme.labelMediumFont(color: VineTheme.primary),
-        ),
-      ),
-    );
-  }
-}
-
-class _VerifiedAccountsSection extends StatelessWidget {
-  const _VerifiedAccountsSection();
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = context.l10n;
-    final claims = context.select<MyProfileBloc, List<IdentityClaim>>((bloc) {
-      final state = bloc.state;
-      if (state is MyProfileLoaded) return state.verifiedClaims;
-      if (state is MyProfileUpdated) return state.verifiedClaims;
-      return const [];
-    });
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(left: 16, bottom: 8),
-            child: Text(
-              l10n.profileEditVerifiedAccountsTitle,
-              style: VineTheme.labelMediumFont(color: VineTheme.lightText),
-            ),
-          ),
-          if (claims.isNotEmpty) ...[
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16),
-              child: VerifiedAccountsRow(claims: claims),
-            ),
-            const SizedBox(height: 8),
-          ],
-          const _GetVerifiedTile(),
-        ],
-      ),
-    );
-  }
-}
-
 typedef VerifierRoutePusher =
     Future<void> Function(String location, {Object? extra});
 
@@ -1920,30 +1401,4 @@ Future<bool> launchVerifierFlow({
     myProfileBloc.add(const MyProfileFetchRequested());
   }
   return launched;
-}
-
-class _GetVerifiedTile extends StatelessWidget {
-  const _GetVerifiedTile();
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = context.l10n;
-    return ListTile(
-      title: Text(
-        l10n.profileEditGetVerifiedCta,
-        style: VineTheme.titleMediumFont(),
-      ),
-      subtitle: Text(
-        l10n.profileEditGetVerifiedSubtitle,
-        style: VineTheme.bodyMediumFont(color: VineTheme.lightText),
-      ),
-      trailing: const DivineIcon(
-        icon: DivineIconName.caretRight,
-        color: VineTheme.lightText,
-      ),
-      onTap: () => context.read<ProfileEditorBloc>().add(
-        const VerifierLaunchRequested(),
-      ),
-    );
-  }
 }
