+++ mobile/lib/screens/profile_setup/view/view.dart
@@ -1,0 +1,2 @@
+export 'profile_setup_screen.dart';
+export 'profile_setup_view.dart';
