+++ mobile/lib/screens/profile_setup/widgets/widgets.dart
@@ -1,0 +1,5 @@
+export 'banner_editing_block.dart';
+export 'profile_setup_upload_errors.dart';
+export 'public_key_link.dart';
+export 'save_button.dart';
+export 'verified_accounts_section.dart';
